Add support for multiple discrete GPIOs and ranges
It is often necessary to examine a number of pins but not all of them,
so add the ability to use comma-separated lists of individual GPIOs
or ranges (x-y), e.g. "7,9-11", where before a single GPIO could be
used.

Take the opportunity to use a few more symbolic constants and have
a light spring-clean.

Signed-off-by: Phil Elwell <phil@raspberrypi.org>